+++ Raveenthiran.com/raveenthiran-obscura/functions.php
@@ -6,7 +6,7 @@
 
 if ( ! defined( 'ABSPATH' ) ) exit;
 
-define( 'NR_THEME_VERSION', '4.41.0' );
+define( 'NR_THEME_VERSION', '4.42.0' );
 
 /* ─────────────────────────────────────────────────────────────
  * Setup